WIP: Fix workspace ingress to work direct links without trailing slash
Add an NginX snippet to the instances ingress:
When the URL does not request a file and does not end with a `/`,
redirect to add a trailing `/`.
This allows correct resolution of relative paths to resources (e.g. js, css files).

TODO: works without keycloak; ends in redirect loop with keycloak

WillFix #23

Contributed on behalf of STMicroelectronics and CEA

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>