Storage class and amount configuration for workspace volumes
This changes the default persistent volume creation to use a storage class.
Before, this was only done when configuring the GKE cloud provider.
Now, the K8S cloud provider behaves this way and an additional cloud provider
MINIKUBE was added to use the former behavior of the K8S provider.
With this, the GKE cloud provider is obsolete and was deleted.

In addition, the used storage class and the amount of requested storage
can now be configured via helm values.

- Add arguments `--storageClassName` and `--requestedStorage` to the operator
- Make persistent storage claiming via storage class the default
- Use the former default - explicitly creating the persistent volume - for
  new cloud provider MINIKUBE only
- Delete cloud provider GKE
- Rename files to fit the new default and cloud provider
- Adapt values files
- Extract persistent volume template replacements to injectable class
  and bind it in the theia cloud DI module.
- Create reusable TemplateReplacements interface

Part of #92

Contributed on behalf of STMicroelectronics

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>