Reduce log level for app definitions allowing indefinite instances
Reduce log level from info to debug because this is the expected result
directly following from the app definition.
With info, this polluted the log with the same information over and over
when a session of an app definition with indefinite max instances is started.

Contributed on behalf of STMicroelectronics

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>