Add monitor theia extension
Added `monitor-theia` extension.
Works the same way as the monitor vscode extension.
Adjusted the way ports are replaced in deployment/service.
This allows to specify the same port for the monitor and the app itself.
Also make sure that the monitor not being enabled does not crash the operator.

Contributed on behalf of STMicroelectronics.