Allow user-scoped requests only for the same authenticated user
So far, most methods did not verify that the user named in the request
is the same as the authenticated user.

This commit fixes this by overloading the evaluateRequest method in
BaseResource with a second version taking a UserScopedServiceRequest.

For these requests, request.user must either be the same as the
authenticated user's identifier or be omitted. In the latter case, the
user's identifier is used automatically.

In anonymous mode, the request.user property must be set because there
is no other way to determine the user.

In both cases, the resolved user identifier to use is returned as part
of the new type `EvaluatedRequest`.