#200: Initial crash loop prevention for custom resources
- Extend the resource clients for app definitions, sessions, and
workspaces to be able to handle the status subresource.
- Add basic status for our CRDs to track handling state of the operator
- Prevent crashloops by never handling resources that are in error or
handling state. The latter indicates an unexcepted crash during handling
(e.g. NPE).
- NOTE: The previous step was only dony for LAZY handlers. EAGER
handlers did not change
- Exemplary sub steps for volume claim an attachment for workspace
resources
- Increase resource versions for all CRs
- Move API, KIND, and CRD_NAME constants from Spec to resource classes
(e.g. from SessionSpec to Session) because they belong to the resource
itself. This became apparent when adding the Status to the resources

Note: Status handling can probably be refactored to be a bit cleaner and
make use of better code-reuse.

Part of #200