#200: Initial crash loop prevention for custom resources (#216)
NOTE: Status handling for app definitions is commented out as there are still some issues

- Catch all exceptions when adding session, workspace, appdefinition resources and add error state
- Extend the resource clients for app definitions, sessions, and
workspaces to be able to handle the status subresource.
- Add basic status for our CRDs to track handling state of the operator
- Prevent crashloops by never handling resources that are in error or
handling state. The latter indicates an unexcepted crash during handling
(e.g. NPE).
- NOTE: The previous step was only dony for LAZY handlers. EAGER handlers did not change
- Exemplary sub steps for volume claim an attachment for workspace resources
- Increase resource versions for all CRs
- Move API, KIND, and CRD_NAME constants from Spec to resource classes
(e.g. from SessionSpec to Session) because they belong to the resource
itself. This became apparent when adding the Status to the resources

Part of #200