#200: Set CRs to error status if handling was interrupted before
When handling of a resource is unexpectedly interrupted without throwing
an error, it currently remains in HANDLING state and is never handled
again. This could happen if the operator is shut down mid handling, e.g.
because of resource constraints.

This commit sets the resource to ERROR state and adds a message that
handling was unexpectedly interrupted before.

Contributed on behalf of STMicroelectronics

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>