Provide conversion webhook from AppDefinition v7 to v8
Rename AppDefinition versions from vXbeta to v1betaX.
This ensures that kubectl will properly detect the newest version.
Provide Hub for AppDefinitions - this is a subset of all fields across versions.
The hub provides a constructor that transform AppDefinitions v7 and v8 to the hub.
The AppDefinition v8 and v7 provide a constructor that transforms from the hub to the version.
Add mappers for the two versions. The mapper calls the constructors.
The mappers are registered to the conversion controller.
New AppDefinition version need to:
- provide a constructor to be created from the hub
- deprecate the old version
- update the constructors in the hub
- add a mapper to the conversion controller

Contributed on behalf of STMicroelectronics