Update CI to new versions and automate publishing
Provide reuseable workflows for publishing npm packages and docker images.
Both will:
- Build on PRs for testing.
- Publish under the next tag and a specific sha version on main merges.
- Publish the version and the latest tag when a releases is published via GH.
Note that the next publishing only happens for versions that end with -next.
This way release versions (e.g. 0.9.0) are not automatically published
when merged to master, instead a release is required.
This allows easy creation of patch versions with out merging them on main.

Each component has its own workflow that calls the reuseable workflows.
These workflows are only triggered when files of that component changed.

Contributed on behalf of STMicroelectronics