Rename publish npm script to avoid rerun of publish
Renamed publish script to publish:latest.
Before the script was executed everytime we ran npm publish.
This then fails because you cannot publish the same version twice.
Also prevents the latest tag to be applied on next publish.
Also it probably would have ended in a loop.

Contributed on behalf of STMicroelectronics