Do not build on full release main pushes
Introduce step that checks the version so we can use the information in the IF statment of jobs.
Unfortunately ENVs are not available there, hence the workaround.
Queue monitor-vscode after theia-cloud-demo so they can be build together with one run.
Automate the detection of the version for the demo flow.
Now there is no need to specify on how they should be published.

Contributed on behalf of STMicroelectronics