terraform: Use bitnami legacy images to fix keycloak setup
Bitnami has removed their images in the dockerhub bitnami repository in favor of a paid service.

This uses the bitnamilegacy repository containing the old images.
They are no longer updated but the Keycloak Helm chart used by us is out of date anyway. Thus, the images being used are effectively the same.