fix: Remove ingress rules for all hosts on session deletion
Previously, the LazySessionHandler only cleaned up a session's ingress rule for the main instances host. However, sessions may define a list of subdomains of the main instances host that are also routed to the session (e.g. to support webviews).
These rules were never cleaned up.

With this, the ingress rules for all relevant hosts are removed.
In addition, logging output and error handling for these cleanups is improved.

Fixes #453